feat: export gamma computation util
