+++ src/entities/DopplerFactory.ts
@@ -50,7 +50,7 @@ import {
   DEFAULT_V4_INITIAL_PROPOSAL_THRESHOLD,
   DEFAULT_CREATE_GAS_LIMIT,
 } from '../constants'
-import { MIN_TICK, MAX_TICK, isToken0Expected } from '../utils'
+import { computeOptimalGamma, MIN_TICK, MAX_TICK, isToken0Expected } from '../utils'
 import { airlockAbi, bundlerAbi, DERC20Bytecode, DopplerBytecode, DopplerDN404Bytecode } from '../abis'
 
 // Type definition for the custom migration encoder function
@@ -485,7 +485,7 @@ export class DopplerFactory<C extends SupportedChainId = SupportedChainId> {
     const addresses = getAddresses(this.chainId)
 
     // 1. Calculate gamma if not provided
-    const gamma = params.auction.gamma ?? this.computeOptimalGamma(
+    const gamma = params.auction.gamma ?? computeOptimalGamma(
       params.auction.startTick,
       params.auction.endTick,
       params.auction.duration,
@@ -1465,33 +1465,7 @@ export class DopplerFactory<C extends SupportedChainId = SupportedChainId> {
   }
 
   // computeTicks moved to builders. No longer needed here.
-
-  /**
-   * Compute optimal gamma parameter based on price range and time parameters
-   * Gamma determines how much the price can move per epoch during the sale.
-   */
-  private computeOptimalGamma(
-    startTick: number,
-    endTick: number,
-    durationDays: number,
-    epochLength: number,
-    tickSpacing: number
-  ): number {
-    // Calculate total number of epochs
-    const totalEpochs = (durationDays * DAY_SECONDS) / epochLength
-    const tickDelta = Math.abs(endTick - startTick)
-    // Base per-epoch movement in ticks
-    let perEpochTicks = Math.ceil(tickDelta / totalEpochs)
-    // Quantize up to the nearest multiple of tickSpacing
-    const multiples = Math.ceil(perEpochTicks / tickSpacing)
-    let gamma = multiples * tickSpacing
-    // Ensure minimum of one tickSpacing
-    gamma = Math.max(tickSpacing, gamma)
-    if (gamma % tickSpacing !== 0) {
-      throw new Error('Computed gamma must be divisible by tick spacing')
-    }
-    return gamma
-  }
+  // computeOptimalGamma moved to utils.
 
   // -----------------------------
   // Bundler helpers (Static/V3)
